requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery<5.0
certifi==2023.7.22
chardet==3.0.4
Django==1.8.2
django-haystack==2.7.0
django-redis==3.8.4
django-redis-sessions==0.5.6
django-tinymce==2.9.0
fdfs-client-py3==1.2.6
idna==2.8
itsdangerous==1.1.0
jieba==0.39
kombu==3.0.37
mutagen==1.42.0
Pillow==9.0.1
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.0
pytz==2019.1
redis==3.2.0
redis-py-cluster==1.3.6
requests==2.31.0
sqlparse==0.3.0
urllib3==1.26.5
uWSGI==2.0.18
vine==1.3.0
Whoosh==2.7.4
